+++ jdaviz/configs/imviz/tests/test_tools.py
@@ -65,43 +65,6 @@ class TestPanZoomTools(BaseImviz_WCS_WCS):
         t_linkedpan.deactivate()
 
 
-# We use a new test class to avoid a dirty state from previous test.
-class TestSinglePixelRegion(BaseImviz_WCS_WCS):
-    def test_singlepixelregion(self):
-        # NOTE: This only works in pixel linking now, not WCS linking.
-
-        t = self.imviz.default_viewer._obj.toolbar.tools['jdaviz:singlepixelregion']
-        t.activate()
-
-        # Create a region while viewing reference data.
-        t.on_mouse_event({'event': 'click', 'altKey': False, 'domain': {'x': 1, 'y': 2}})
-        regions = self.imviz.get_interactive_regions()
-        assert len(regions) == 1
-        reg = regions['Subset 1']
-        assert (isinstance(reg, RectanglePixelRegion) and reg.center.x == 1 and reg.center.y == 2
-                and reg.width == 1 and reg.height == 1)
-
-        # Clicking again will move the region, not creating a new one.
-        t.on_mouse_event({'event': 'click', 'altKey': False, 'domain': {'x': 2, 'y': 3}})
-        regions = self.imviz.get_interactive_regions()
-        assert len(regions) == 1
-        reg = regions['Subset 1']
-        assert (isinstance(reg, RectanglePixelRegion) and reg.center.x == 2 and reg.center.y == 3
-                and reg.width == 1 and reg.height == 1)
-
-        # Create a new region while viewing dithered data.
-        # Region will still be w.r.t. reference data, that is, x and y from domain stay the same.
-        self.imviz.default_viewer.blink_once()
-        t.on_mouse_event({'event': 'click', 'altKey': True, 'domain': {'x': 3, 'y': 4}})
-        regions = self.imviz.get_interactive_regions()
-        assert len(regions) == 2
-        reg = regions['Subset 2']
-        assert (isinstance(reg, RectanglePixelRegion) and reg.center.x == 3 and reg.center.y == 4
-                and reg.width == 1 and reg.height == 1)
-
-        t.deactivate()
-
-
 def test_blink(imviz_helper):
     viewer = imviz_helper.default_viewer._obj